Add basic example of group selector from .blend
Added two ways, through enum or through collection.

Enum way is simpler, but needs additional caching, and afaik is not
serachable in the UI.

Collection way is a bit more tedious, but stores its own data itself,
and is searcheable with prop_search() layout function.

Note that both are not linked together (setting enum version does not
alter collection choice and vice-versa).